### 2025-04-11 #698 @Visorbearer
PFW: Tools for Filtering and Processing Data from Project FeederWatch
Submitting Author Name: Mason W. Maron
Submitting Author Github Handle: @Visorbearer
Other Package Authors Github handles: (comma separated, delete if none)
Repository:  https://github.com/Visorbearer/PFW
Version submitted: 0.0.1
Submission type: Standard

Version accepted: TBD

---

```
Package: PFW
Title: Tools for Filtering and Processing Data from Project FeederWatch
Version: 0.0.1
    person("Mason", "Maron", "W.", "mmaron101@gmail.com", role = c("aut", "cre"))
Description: Provides tools to import, clean, filter, and prepare Project
    FeederWatch data for analysis. Includes functions for taxonomic
    rollup, easy filtering, zerofilling, merging in site metadata, and more.
    rvest,
    utils
    withr,
```

The package is based around downloading and importing data from Project FeederWatch and then streamlining handling, filtering, and processing
of said data.

The target audience is researchers who wish to use Project FeederWatch data for scientific analysis. Project FeederWatch, run by
Birds Canada and the Cornell Lab of Ornithology, is an extensive dataset of bird records from various sites,
but typically backyards, from across North America. This package streamlines, optimizes, and simplifies the data preparation stage
of analysis with this dataset, a much needed tool as the data is often misinterpreted or prepared incorrectly.

No currently existing R packages serve this function. Although the filtering could be done in a longer and more complex
process using filtering packages like `dplyr`, doing so is often what results in FeederWatch data misinterpretation or mishandling.

N/A
