### 2025-04-14 #698 @maurolepore
@Visorbearer thanks a lot for sharing your work with rOpenSci, and for responding quickly to the bot's feedback.

I'll announce it to the editorial board after a few preliminary checks which I'll post here ASAP.
