### 2025-04-17 #698 @maurolepore
Here are my preliminary checks. They are mostly to facilitate conversations with the editorial board. If any of this seems helpful, now it may be a good time to make changes.

### Preliminary checks:

    > In README I see a good description of what the package does but "the case for it" is more about *why*. What motivated you to create this package? What was hard to do before that is easier now?
    > I see `update_taxonomy()` is the only function without the `pwf_*` prefix. Is this intentional?
    > In README I see a repetition of what's in "Reference". This doesn't seem to add much value and can get tricky to maintain.
    > In README and elseswhere I did not see the actual output of the code. I wonder if there is a specific reason this is the case? If possible that would greatly help reviewers and users to understand what to expect without having to run the code.
  > I do see the fit for data retrieval but not so much for data munging. For convenience I paste the definition here. Can you please review and explain if/how it fits?:
 > data munging: Packages for processing data from formats above. This area does not include broad data manipulations tools such as reshape2 or tidyr, or tools for extracting data from R code itself. Rather, it focuses on tools for handling data in specific scientific formats generated from scientific workflows or exported from scientific instruments. (Examples: plateR for reading in data structured as plate maps for scientific instruments, or phonfieldwork for processing annotated audio files for phonics research)

  > I didn't dig deep enough to see how it handles HTTP requests in tricky situations, e.g. with no internet. Can you please briefly explain in the context of this section of our guide?: https://devguide.ropensci.org/building.html#testing
---

* I see the bot highlights `pwf_zerofill()` as pretty complex. Is there a way to simplify this logic? Do the tests cover all these many branches of the logic?
* A quick look at `pwd_download()` suggests it likely does more than one just download data?

---
