### 2025-04-17 #698 @Visorbearer
@maurolepore Thanks for those checks! Here are my responses:

> "In README I see a good description of what the package does but "the case for it" is more about why. What motivated you to create this package? What was hard to do before that is easier now?"

Project FeederWatch data is easy to access, but can often be tricky to work with. I myself have made custom workflows to process it for a research paper (the workflows that inspired this package), but felt that independent functions which greatly simplify the experience would benefit researchers, especially those less interested in doing a deep dive into the dataset's components. Similar packages exist for eBird (`auk`) and NestWatch (`nestwatchR`), but no package yet exists for Project FeederWatch in this context. Additionally, in personal communications with Project FeederWatch staff at Cornell, I was told they "regularly see folks misinterpreting/misusing [the] open access datasets, and want to help the research community avoid the common pitfalls", which this package aims to achieve.

> “I see update_taxonomy() is the only function without the pwf_* prefix. Is this intentional?”

It’s intentional in that I felt “pfw_update_taxonomy()” was too long of a function name to call, but I can see why attaching the pfw_ prefix would be reasonable and am happy to change it if that is the best practice.

> “In README I see a repetition of what's in "Reference". This doesn't seem to add much value and can get tricky to maintain.”

That’s fair-- If I remember correctly, I added that section as I set up the README before the pkgdown page, but will remove it from the README shortly.

> “In README and elsewhere I did not see the actual output of the code. I wonder if there is a specific reason this is the case? If possible that would greatly help reviewers and users to understand what to expect without having to run the code.”

This is a good call and an oversight on my part; I will update the README to include some example outputs. Thanks for the suggestion!

> I do see the fit for data retrieval but not so much for data munging. For convenience I paste the definition here. Can you please review and explain if/how it fits?: “data munging: Packages for processing data from formats above. This area does not include broad data manipulations tools such as reshape2 or tidyr, or tools for extracting data from R code itself. Rather, it focuses on tools for handling data in specific scientific formats generated from scientific workflows or exported from scientific instruments. (Examples: plateR for reading in data structured as plate maps for scientific instruments, or phonfieldwork for processing annotated audio files for phonics research)”

Project FeederWatch data is available in a specific format provided by the source, including multiple separate files with different (but related) data. The data retrieval side of things is pulling that data from Project FeederWatch-- data munging comes in with the processing and filtering of that data within the R environment throughout the workflow, working with the data in the base format in a more readable/accessible way for the user.

> I didn't dig deep enough to see how it handles HTTP requests in tricky situations, e.g. with no internet. Can you please briefly explain in the context of this section of our guide?: https://devguide.ropensci.org/building.html#testing

This is a good catch. I was missing stop messages if there was no internet connection. I’ve added those in the relevant functions. I also added `skip_on_cran()` for the tests that specifically call downloads from the webpages. The actual web-accessing functions are pretty minimal and only used for file downloads, so curl’s `has_internet()` should cover them.

> I see the bot highlights pwf_zerofill() as pretty complex. Is there a way to simplify this logic? Do the tests cover all these many branches of the logic?

The logic is a bit complicated because of the nature of zerofilling-- you have to reapply filters to the original dataset (except species filters) and then use a `SUB_ID` x `SPECIES_CODE` combination to derive zeros for all species in the input data (which is more complex with more than one species). As it stands, I don’t think the script is overwhelmingly complex itself, but it may be possible to streamline in ways that I am not seeing. Any feedback I receive on ways to abbreviate the function will, of course, be greatly appreciated :)

> A quick look at pwd_download() suggests it likely does more than one just download data?

It technically does slightly more, but still is just a data download function. Project FeederWatch data downloads as individual .csv files within .zip files, so it downloads the .zips, unzips them to get the .csvs, and then removes the original .zip files. I think this will probably be fine as-is.
