### 2025-05-06 #698 @maelle
---

Thank you for your submission! :bird:

I very much enjoyed this code line: https://github.com/Visorbearer/PFW/blob/ff66a30e3e00a3aed11bd42a0f17425de0c8c6f1/R/check_taxonomy.R#L46 :grin:

I'm happy to discuss any of the comments.

##### Toolset: HTTP package and caching

In this package, to download data you use rvest and utils.
I would recommend instead to switch to httr2 (or curl, or crul) because the rvest call are not too complex, and because it would make it more straightforward to add tests.

``` r
page <- httr2::request("https://feederwatch.org/explore/raw-dataset-requests/") |>
  httr2::req_user_agent("PFW package https etc etc") |>
  httr2::req_perform() |>
  httr2::resp_body_html()
page |>
  xml2::xml_find_all(".//a[contains(@href, '.zip')]") # you might want a more specific regex
#> {xml_nodeset (8)}
#> [1] <a href="https://cdn.feederwatch.org/data/202406/PFW_all_1988_1995_May202 ...
#> [2] <a href="https://cdn.feederwatch.org/data/202406/PFW_all_1996_2000_May202 ...
#> [3] <a href="https://cdn.feederwatch.org/data/202406/PFW_all_2001_2005_May202 ...
#> [4] <a href="https://cdn.feederwatch.org/data/202406/PFW_all_2006_2010_May202 ...
#> [5] <a href="https://cdn.feederwatch.org/data/202406/PFW_all_2011_2015_May202 ...
#> [6] <a href="https://cdn.feederwatch.org/data/202406/PFW_all_2016_2020_May202 ...
#> [7] <a href="https://cdn.feederwatch.org/data/202406/PFW_all_2021_2024_May202 ...
#> [8] <a class="btn arrowed" href="https://feederwatch.org/wp-content/uploads/2 ...
```

Some of the files you download are not updated often.
To know when to update them, you could rely on caching.
See for instance https://httr2.r-lib.org/reference/req_cache.html

I tried doing a HEAD request to one of the file, it does include a `last-modified` header.

``` r
httr2::request("https://cdn.feederwatch.org/data/202406/PFW_all_1988_1995_May2024_Public.csv.zip") |>
  httr2::req_method("HEAD") |>
  httr2::req_perform() |>
  httr2::resp_headers()
#> <httr2_headers>
#> content-type: application/zip
#> content-length: 28939188
#> date: Mon, 05 May 2025 14:37:23 GMT
#> last-modified: Tue, 04 Jun 2024 03:17:19 GMT
#> etag: "498c63dd586b75222d5a13eae63c07ed-4"
#> x-amz-server-side-encryption: AES256
#> accept-ranges: bytes
#> server: AmazonS3
#> x-cache: Hit from cloudfront
#> via: 1.1 e3dbec42cd15573654eebd0dfddc01fe.cloudfront.net (CloudFront)
#> x-amz-cf-pop: MRS52-P5
#> x-amz-cf-id: 8dcbWio9QZxNY18cEbK6nnbvNDU-W2S9RSee5YN9sGWeq_nXprvByA==
#> age: 62255
```

So you could update the file as necessary, relying on `httr2::req_cache()` or doing yourself a HEAD request then a GET request only if necessary.

On that topic, the files that are downloaded should by default go to an *user directory* rather than the package's installed version.
Since your package depends on R 4.1.0 you're in luck as `tools::R_user_dir()` was added in R 4.0.0 (otherwise you'd have had to use the [backports package](https://github.com/r-lib/backports) or the [rappdirs package](https://rappdirs.r-lib.org/)).
User directories are special directories on one's machine that are [standard locations](https://blog.r-hub.io/2020/03/12/user-preferences/#preface-standard-locations-on-the-users-machine) to store user's configuration, stored data.

``` r
tools::R_user_dir("PFW", which = "data")
#> [1] "/home/maelle/.local/share/R/PFW"
tools::R_user_dir("PFW", which = "cache")
#> [1] "/home/maelle/.cache/R/PFW"
```

For testing of HTTP requests, you could refer to [HTTP testing in R](https://books.ropensci.org/http-testing) but it is outdated as regards [vcr's development version](https://github.com/ropensci/vcr) (version that looks great).
Now some of the files you download are huge so you might need to manually edit recorded versions to make them smaller.

I don't understand the need for https://github.com/Visorbearer/PFW/blob/ff66a30e3e00a3aed11bd42a0f17425de0c8c6f1/R/global_storage.R#L11 -- can you please explain?

##### Documentation

- I'd recommend using `README.Rmd` instead of just `README.md` as it'll be a more natural way to show, and update, code output. `usethis::use_readme_rmd()` will create the README for you, then inside you can port what you currently have in `README.md` and make the code chunk an actual executable code chunk. `devtools::build_readme()` will knit `README.Rmd` using a temporary installation of the development version of the package.

- If you use pkgdown's Bootstrap 5 templates, the website will look better. In particular the font will be bigger and it'll gain search. You only need to add those lines in `_pkgdown.yml`:

```yml
template:
  bootstrap: 5
```

- The navbar configuration in `_pkgdown.yml` is not useful as far as I can tell. pkgdown will automatically create a "Get Started" link to the vignette [since it has the same name as the package](https://pkgdown.r-lib.org/articles/pkgdown.html#articles).

- I see you put backticks around your package's name in the README. You can of course keep doing that, but I think I more often see the package name written as normal text.

- In the README instead of devtools you could recommend using pak that is more prevalent these days or at least becoming more prevalent (for instance that's what documented in usethis' README).

- I wonder whether the README could feature a plot, a map? Not a complex one, but an illustration.

- In the README you mention using one's own data. To me it is a bit confusing as I thought the package was primarily about downloading and using data from Project Feeder Watch.

- The examples should be executable by users. Many of them are currently wrapped in `\dontrun{}` and refer to data that hasn't been created like in `data_2001 <- pfw_date(data, year = 2001)`. Examples should be self-contained. They could use the example dataset. For the functions downloading data, you could use [`#' @examplesIf interactive()`](https://roxygen2.r-lib.org/articles/rd.html?q=examplesIf#examples) to have them executed only in interactive sessions (so not in CRAN checks for instance).

- In DESCRIPTION you should add your [ORCID ID](https://ropensci.org/blog/2018/10/08/orcid/).

##### Dependencies

Beside replacing rvest, I wonder whether you could consolidate the dependency list by using  _either_ dplyr or data.table?
Is there a special reason to use both?

Your package depends on R 4.1.0 so you don't need `stringsAsFactors = FALSE` that became the [default in R 4.0.0](https://blog.r-project.org/2020/02/16/stringsasfactors/).

##### Exported functions

Why export `load_region_lookup()` and `check_taxonomy()`, would users call them directly?
If not, I'd recommend not exporting them.
To keep using Rd tags to document them you could either

- use `#' @NoRd` which means no manual page is generated.
- use the [devtag](https://github.com/moodymudskipper/devtag) package with which manual pages are generated, but not for the installed version of the package.

##### Installed data

In

https://github.com/VisorBearer/PFW/blob/ff66a30e3e00a3aed11bd42a0f17425de0c8c6f1/R/load_region_lookup.R#L13-L15

Why would the data be missing at all?

I have the same question about `pfw_example()`. :smile_cat:

##### nocov

- Why wrap `pfw_example()` in nocov tags, couldn't it be tested, isn't it used in tests anyway?

- Why is this code part not tested:

https://github.com/VisorBearer/PFW/blob/ff66a30e3e00a3aed11bd42a0f17425de0c8c6f1/R/pfw_filter.R#L56-L62

- For this code could you create fake wrong data?

https://github.com/VisorBearer/PFW/blob/ff66a30e3e00a3aed11bd42a0f17425de0c8c6f1/R/pfw_import.R#L45-L48

- Same question for:

https://github.com/VisorBearer/PFW/blob/ff66a30e3e00a3aed11bd42a0f17425de0c8c6f1/R/pfw_import.R#L45-L48

- I'd recommend going through all nocov tags to ensure the code that's "escaped" can't be tested. Of course if testing it requires a huge setup and it's obvious code, it can be ok not to test some parts.

##### Tests

- In test files such as https://github.com/Visorbearer/PFW/blob/ff66a30e3e00a3aed11bd42a0f17425de0c8c6f1/tests/testthat/test-pfw_download.R#L7 you could instead simply use `withr::local_tempdir()`. It will clean after itself so you could avoid the `defer()` and `unlink()` code.

- This test code

https://github.com/Visorbearer/PFW/blob/ff66a30e3e00a3aed11bd42a0f17425de0c8c6f1/tests/testthat/test-check_taxonomy.R#L3C1-L9C4

could be simpler to read as (using `local_` instead of `with_`, and relying on [snaphsot tests](https://testthat.r-lib.org/articles/snapshotting.html)) but that might be personal preference:

```r
  withr::local_envvar(PFW_TRANSLATION_DIR = tempfile())
  expect_snapshot(check_taxonomy())
```

##### Argument checks

For argument checks like this one:

https://github.com/VisorBearer/PFW/blob/ff66a30e3e00a3aed11bd42a0f17425de0c8c6f1/R/pfw_rollup.R#L16-L19

- you could rely on R's error for when the user provides no argument as it's quite clear and usual. It would look like so:

```r
Error in `pfw_rollup()`:
! argument "data" is missing, with no default
```

- you could check more specifically for the class or type of the input. Example from the igraph package: https://github.com/igraph/rigraph/blob/9a2816248eb837077b5af28bad7b1b9c0713b128/R/hrg.R#L316 (using the [cli package](https://blog.r-hub.io/2023/11/30/cliff-notes-about-cli/) which is quite near but not compulsory)

##### Data loading

I very rarely see `rm()` and `gc()` in code.
Therefore I wonder whether the code around https://github.com/Visorbearer/PFW/blob/ff66a30e3e00a3aed11bd42a0f17425de0c8c6f1/R/pfw_import.R#L51 could be refactored.
For instance you could create a helper function that reads and checks the file, returns either the data or `NULL`.
Then you'd apply the function to files using an apply function (or `purrr::map()`) then append all the results together.
The [functional chapters of Advanced R by Hadley Wickham](https://adv-r.hadley.nz/fp.html) might be an useful read.
The principles apply (unintended pun!) for the apply family of functions and purrr alike.

---
