### 2025-05-19 #698 @Visorbearer
Hi @maelle,

I believe I've addressed everything below and in the most recent commit to GitHub, but please let me know if I missed anything! Thank you for all the feedback!

### Toolset: HTTP package and caching

I've:
- Replaced `rvest` use with `httr2` as suggested
- Correctly sent downloaded files to a user directory for `pfw_download()` and `pfw_sitedata()`
- Updated testing alongside the updated functions

I am not sure how useful caching would be in this case. I may be misunderstanding how this function works, but it seems like the user would still need to redownload the data in order to see if the last-modified date had changed, in which case they may as well just use the newly-downloaded data anyway. If I am misunderstanding your intention there, please let me know!

To answer your question about `global_storage.R`, it's used to create a hidden environment which stores the original data for later zerofilling, as well as to define global variables. There may be a better way to do this (I am still relatively new to this!), so please let me know if so.

### Documentation

I've:
- Updated with a `README.Rmd` as suggested
- Added pkgdown's bootstrap 5 template
- Removed the navbar
- Switched to `pak` instead of `devtools` in the README
- Added a map from *Bonter & Greig 2021* to the README
- Reworded the "your data" comment in the README for clarity
- Added my ORCID to the DESCRIPTION file

I removed /dontrun{} and incorporated examples for most functions. `pfw_zerofill()` could be changed, but it is a very burdensome function to run, especially on the whole example dataset. `pfw_import()` also relies on `pfw_download()`, so it does not really work in a standalone example, I think. I can still edit them if preferred, though.

I also left the backticked `PFW` package name. I like that it sort of sticks out from the text that way, but if it's preferred, I can change that easily.

### Dependencies

I've removed `data.table` use and transitioned all of the relevant code to `dplyr`. I've also removed `stringsAsFactors = FALSE` where it occurred.

### Exported functions

I've used `@NoRd` as suggested and prevented `load_region_lookup()` and `check_taxonomy()` from exporting.

### Installed data

The data shouldn’t be missing for the region lookup, but could be if the user manually deleted it for some reason-- it’s mostly a failsafe. People can do weird things sometimes, so I just wanted to limit potential breaking of the package where I could think of it happening.

The example from `pfw_example()` does not come pre-downloaded since it is too large for CRAN, so it must be downloaded manually with `pfw_example()` the first time.

### nocov

I tried to reduce what was nocov'd extensively as you suggested, though I struggled a bit here. I had to leave some nocov sections where the tests are skipped on CRAN, since that apparently skips them in covr, too. Overall, I tried to leave only reasonable nocov lines.

### Tests

I've revised the test files with `withr::local_tempdir()` as suggested.

### Argument checks

I've adjusted this as suggested in all functions except `pfw_download()` and `pfw_filter()`, where I feel the comments alongside the stops help the user understand what they need to add for the function to work properly.

### Data loading

I've removed the `rm()` and `gc()` calls as suggested.

---

Hopefully, this covers and addresses all of your very helpful input! I should be much more on top of things and responsive now. I look forward to the next steps!
