### 2025-05-19 #698 @Visorbearer
Thanks @maelle! Those are all good inputs.

> Maybe you could set the default to `NULL` and then do the computation inside the function

Done!

> I wonder whether this function should invisibly return the path instead of ending with a message

Updated as well.

> Regarding caching, I mean the package could try and update the data. Maybe in each session? In any case, your code would do a HEAD (no data downloaded, just the headers of the response) not a GET request (or you'd use the caching function in httr2) to see when it was last updated and only do a GET request if necessary. Does that make any sense?

Yes, that makes a lot more sense-- thanks for explaining it further. I understand now, and I've added that into `update_taxonomy()` so that it can let the user know if the species translation table is already up to date before they redownload it.

> You can escape examples so that they are only run in interactive sessions, that might help for long running examples

I've added that for some of the longer examples I left it off of in yesterday's revisions.

> Maybe add the `NOT_CRAN` variable to the GHA workflow?

Done, and that does seem to have resolved the issue.

---

Thanks again for all the feedback!
