### 2025-06-06 #698 @maelle
@Visorbearer I think it's better to wait for @paul-carteron's review indeed.
