### 2025-06-09 #698 @paul-carteron
Estimated hours spent reviewing: 6h30

---

"Bravo" for your package! The goal is clearly stated, and the process to achieve it is easy to follow. Even though I’ve never worked with this type of species dataset before, I found the package intuitive to use. It’s clear that a lot of thought has been put into making things easier for the end user. This is a really well-crafted package.

Below, I’ve included a few suggestions to further improve the user experience and help make long-term maintenance of your package even easier.

#### General Recommendations

Below are some suggestions and good practices that you may consider implementing across your codebase. They aim to further improve readability, maintainability, and robustness:

* Prefer descriptive variable names – Some generic names like "path" or "data" can be ambiguous in larger contexts. Using more self-explanatory names can make your code easier to maintain and understand over time. It’s often said that naming things accounts for 90% of code readability 😄

* When signaling an error using `stop()`, it’s a good practice to set `call. = FALSE` to produce cleaner, more user-friendly error messages. You can read more about this in [Advanced R](https://adv-r.hadley.nz/conditions.html?q=call.%20=%20FALSE#errors-1).

* When using if conditions, introducing intermediate, self-descriptive variables can greatly improve clarity and long-term maintainability. Here's a simple example adapted from your `pwf_import` function:

# Current version
if (!all(expected_cols %in% names(data))) {
  message("Skipping file (incorrect data structure): ", file)
  next
}

# Suggested version
is_valid_pwf_data <- all(expected_cols %in% names(data))
if (!is_valid_pwf_data) {
  message("Skipping file (incorrect data structure): ", file)
  next
}
```
* I see you don't use a lot of `dplyr` functions — if you want, you could remove the dependency entirely by switching to base R. Of course, this is not mandatory 🙂
Specifically:
   * `bind_rows` → `do.call(rbind, list_of_data)`
   * `filter / select` → `subset()`
   * `mutate` → `transform()`
   * `left_join` → `merge()`

### `pfw_import()`:
#### Default dir :

When using `pwf_import("data_raw/")` as shown in the vignette, I encountered the following error:

Erreur dans pfw_import("data_raw/", year = 2002:2003, region = "Washington") :
  The folder 'data_raw/' does not exist. Please create it and download Project FeederWatch CSV files into it.
```

The issue seems to come from a mismatch between the default folder used by `pfw_download()` and the folder expected by `pfw_import()`. Specifically, `pfw_download()` uses:

folder <- file.path(tools::R_user_dir("PFW", "data"), "data-raw")
```

which, in my case, resolves to:

"C:\\Users\\PaulCarteron\\AppData\\Roaming/R/data/R/PFW/data-raw"
```

So when `pfw_import("data_raw/")` tries to locate a relative folder in the working directory, it fails. That said, calling `pfw_import()` with `folder = NULL` works perfectly.

To ensure a smooth user experience when following the vignette, you could consider one of the following:

* Set `folder = NULL` explicitly in the vignette when calling pfw_import(), to default to the correct internal folder used by `pfw_download()` ;
* Alternatively, if encouraging a custom data directory, set the same path consistently in both functions, e.g.:

pfw_download(folder = "my_custom_folder")
pfw_import("my_custom_folder", year = 2002:2003, region = "Washington")
```

This would ensure the vignette runs without error and avoids confusion for first-time users.

#### Coding style proposition

Here's a small suggestion that’s more about coding style than functionality. I tend to lean toward functional programming, so I have an almost allergic reaction when I see a for loop 😄  (I'm working on it — I know loops aren’t evil!). That said, I couldn’t resist trying to rewrite your `pfw_import()` loop in a more functional style.

Of course, this isn’t something you need to change — your current implementation is perfectly valid. This is just another way of approaching the problem that you might find interesting or helpful in some cases.

expected_cols <- c("SPECIES_CODE", "HOW_MANY", "SUB_ID")

read_pwf_data <- function(file) {
  message("Reading: ", file)
  header <- read.csv(file, nrows = 1)

  is_valid_pwf_data <- !is.null(header) && all(expected_cols %in% names(header))
  if (!is_valid_pwf_data){
    message("Skipping (incorrect structure): ", file)
    return(NULL)
  }

  pwf_data <- read.csv(file)
  return(pwf_data)
}

raw_pwf_data <- files |> lapply(read_pwf_data)
raw_pwf_data <- Filter(Negate(is.null), raw_pwf_data)

if (length(raw_pwf_data) == 0) {
  stop("No valid Project FeederWatch data files were found in the folder.", call. = FALSE)
}

pwf_data <- dplyr::bind_rows(data_list) |>
  dplyr::filter(!is.na(HOW_MANY))
```

Nonetheless, whether you choose functional programming or for loops, you can still make a minor enhancement by making the function read only the headers with `header <- read.csv(file, nrows = 1)` instead of the entire file when checking for right colnames. That allow function to fail fast.

#### Caching ?

Since reading data can be quite long, perhaps you could consider caching the result. I heard about the [`memoise` package](https://memoise.r-lib.org/), which could be a nice way to explore that.

### `pfw_example()`:

Because this is example data that will probably not change frequently, you should add it as a dataset following the rules explained in [R Package (2e)](https://r-pkgs.org/data.html). You chose to use `extdata`, which is more appropriate for showing raw datasets. My understanding is that this isn’t really a raw dataset, since `pwf_import()` already performs some data operations.

My go-to would be to use `usethis::use_data_raw()` to create the script to retrieve the example data, and then store it using `usethis::use_data()`.

Doing it this way can potentially lead to CRAN refusing your package if the dataset exceeds 5MB. In your case, this shouldn’t be a problem, since the dataset is 200,712 bytes (~196 KB), as shown by `tools::checkRdaFiles("data/pfw_example.rda")`.

usethis::use_data_raw("pfw_example")

# in data-raw/pfw_example.R
tmp <- tempdir()
pfw_download(tmp, year = 2002:2003)
pfw_example <- pfw_import(tmp, year = 2002:2003, region = "Washington")

# i'm using compress = "xz", following instruction in R Package data - 7.1.1
usethis::use_data(pfw_example, overwrite = TRUE, compress = "xz")
```

### `update_taxonomy()` :

This function is used for user-modifiable data. My understanding is that for this kind of data, the recommended approach is to use `tools::R_user_dir("yourpkg", "data"`).

Also, I noticed that the function downloads "PFW_spp_translation_table.csv" inside a subfolder called "SpeciesTranslationTable". I think you could simplify things by removing that subfolder and writing the file directly to the main directory.

### `pwf_sitedata()` :

Quite an interesting function — it does a lot! I was a bit confused at first because of all the processing involved.

For easier reading and better maintainability, I’d recommend splitting the function into several smaller ones. This would also make it easier to test each step individually. Here's a possible breakdown:

* `pfw_site_data_download()` : download the dataset if it doesn't already exist ;
* `pfw_sitedata_get_url()` : scrape the web page to find the correct URL;
* `pfw_sitedata_merge()` : merge site_data with pfw_data;
* `pfw_site_data()` : keep this function as the wrapper that combines the three above

### `pwf_filter()` :

What a beautiful function! Many small ones are involved in it, which really helps with readability. Also, storing filters in an attribute over time is a great idea — it makes it easy to see what has been applied and where we are in the process!

The `roll_up` argument was a bit hard for me to understand (I’m not a native English speaker). Adding a small example with one species — with and without `roll_up` — could really help users understand its effect.

#### `month` :

It could be interesting to allow users to set months using names or abbreviations, like `month = c("jan", "feb")` or `month = c("january", "february")`.

To do that, you could create a small helper function to clean and standardize the month input, like:

x <- c("Jan", "Febr", "MAR", "may", "december", 12)
clean_month <- function(x) {
  tryCatch({
    # when i have mix character and numeric in a vector it will be character
    # First i'm trying to coerce x as numeric (ex : "12" to 12), if it works i return number
    num <- suppressWarnings(as.numeric(x))
    if (is.na(num)) stop("not numeric")
    num
    # if it fail, it's probably character month so i use lubridate to parse mont character to number
  }, error = function(e) {
    month(parse_date_time2(x, orders = "B"))
  })
}

res <- sapply(x, clean_month, USE.NAMES = FALSE) |> unique()

```

Also, the documentation can be a bit misleading when using non-sequential or reversed months (a slightly odd case, i agree 😄).

For example, if I want to select January and April, `c(1, 4)` works fine — but `c(4, 1)` returns:

c(4, 5, 6, 7, 8, 9, 10, 11, 12, 1)
```

This happens because `c(4, 1)` is interpreted as a wrapped range, similar to `4:1`. But a user providing `c(4, 1)` probably expects just April and January — especially if the month vector comes from a pipeline and isn’t ordered.

So maybe it’s worth either:

* Clarifying this behavior in the documentation ;
* Modifying the code to allow unordered or reversed month inputs to be interpreted literally.

1:3 # "jan", "feb", "mar
3:1 # "mar", "apr", "may", "jun", "jul", "aug", "sep", "oct", "nov", "dec", "jan"
c(1, 2, 3) # "jan", "feb", "mar
C(3, 2, 1) # "mar", "feb", "jan"
```

#### `species` :

I expect users (especially me 😄) might not know the exact name of the bird. Maybe adding a helper function to search for bird names could help them find the correct one more easily? Something like:

find_species <- function(
    pattern,
    search_in = c("american_english_name", "scientific_name", "species_code"),
    species_table){

  search_in <- match.arg(search_in, c("american_english_name", "scientific_name", "species_code"), several.ok = TRUE)

  matching_row <- sapply(search_in, function(x) {grep(pattern, species_table[[x]], ignore.case = TRUE)}) |>
    unlist() |>
    unique()

  return(species_table[matching_row,])
}

pattern <- "red"
red_species_in_code <- find_species(pattern, search_in = "species_code", species_table)
red_species_in_english_name <- find_species(pattern, search_in = "american_english_name", species_table)
all_red_species <- find_species(pattern, species_table)

# You can then use this in your workflow like below
red_bird_data <- pfw_import(filter = TRUE, species = all_red_species$species_code)

```
Also, as @SunnyTseng recommended, it might make more sense to store it as an internal dataset in the package (in `.rda` format).

However, if you want to implement the functionality mentioned above (e.g. letting users query or inspect the species table), then it would make more sense to store it as an external dataset — still in `.rda` format, but accessible via `PFW::species_table`.

### `region` :

I see you added a dataset for regions that's accessed with `load_region_lookup()`. Another  approach would be to store this data as an `.rda` file using `usethis::use_data()`. That way, users can access it directly with `pfw::region_lookup` or `data("region_lookup")`, which makes it more transparent and easier to inspect.

Also, I’m not sure where the original `region.csv` came from, but a best practice would be to create a script that retrieves and formats this data, and place it in the `data-raw/` folder. This improves reproducibility and makes the data source and transformation steps clear.

### `pfw_dictionary()` :

This is so cool!!

I see you use the data "FeederWatch_Data_Dictionary". If you manually built it, keeping it as a `.csv` makes sense — it’ll be easier to update descriptions later. But if you downloaded it as raw data from a web page, I’d recommend storing it as an `.rda` file using `usethis::use_data()`, and putting the script you used to generate it in the data-raw/ folder.

Also — the fact that users can specify variable names is even cooler!
