### 2025-06-11 #698 @maelle
@Visorbearer yes, working on changes and then writing a response (one issue comment with the response to the two reviews) https://devguide.ropensci.org/softwarereview_author.html#the-review-process, then after posting the response "submitting" it to our records using the bot command https://devguide.ropensci.org/bot_cheatsheet.html#submit-response-to-reviewers

However it's also fine to ask a question to a reviewer in an issue comment before your response if needed! 😸
