### 2025-06-17 #698 @Visorbearer
# Reviewer Comments and Responses/Adjustments
---
### Thank you to @SunnyTseng and @paul-carteron for your reviews (and thanks to @maelle, of course). I appreciate all the feedback greatly and have worked through everything to make a lot of the suggested improvements. Below are my responses to all of the suggestions made as I understood them, so please correct me if I misinterpreted any. If I didn’t make a change, I provided a comment explaining my justification, but am always happy to discuss further and be persuaded :) Without further ado:
---
> Include a flowchart in the README to illustrate the general step-by-step workflow of the package

I really appreciate this suggestion, and am more than happy to change my decision and add it if it feels especially warranted. That said, I do feel that the existing, linear workflow of the README quick start section fills the same niche a flowchart would, making the latter somewhat redundant.
> Consider using the checkmate package instead of if checks for argument checks

I personally find the if checks to be easier for me to understand, especially in looking back on my own code, but am happy to change it if there is an overwhelming preference for something like checkmate instead.
> In pfw_download(), clarify where the data-raw folder is created within the roxygen documentation

Done!
> In the roxygen documentation for pfw_download(), it would be helpful to know the years of data that can be downloaded.

Fixed!
> Add sentences like “data is available from 1988 to the current year” in the roxygen documentation.

Done!
> For pfw_import(), include information in the roxygen documentation that the data can also be downloaded using the pfw_download() function or PFW website before running pfw_import().

Done!
> Edit pfw_example() such that the example is a dataset prepackaged within PFW and pfw_example only loads it into the active environment

Done!
> Consider making the species translation table a dataset in the package instead of a .csv

The species translation table is updated every year by Cornell when they do their annual taxonomy updates and has to be redownloaded by the user through update_taxonomy() if their package version has not been updated. I believe this means it’s easier to keep it as a .csv, but please correct me if I’m wrong!
> Update pfw_download() description to say local directory instead of working directory in the vignette

Fixed!
> Clarify pfw_download code so that it only gives the overwrite prompt if downloading a file already in the folder rather than if there are any files in the folder at all

Fixed!
> Consider more self-explanatory names than ‘path’ or ‘data’ where applicable to remedy ambiguity

In larger functions, I do use descriptive variable names (e.g. existing_files, csv_filenames, downloaded_years in `pfw_download()`). Where I do use path or data, the meaning is typically unambiguous or explained through comments, but I’ve made some revisions where it could still be ambiguous.
> Set call. = FALSE when using stop()

Done!
> Consider adding self-descriptive variables in if checks for long-term clarity on if check function

This is reasonable and a good suggestion, so I have made a change, but have done so by adding descriptive comments instead of additional variables out of personal preference. Please let me know if this was not what you meant!
> Consider removing dplyr from the package requirements since it is minimally used and switch fully to base R

While a good suggestion, I would prefer to keep dplyr in (even in its limited use), as it’s easier for me to read/refer back to and will likely still be used in my future maintenance as well.
> Fix pfw_import()-pfw_download() folder mismatch

Fixed!
> Consider switching out pfw_import() for loop with functional programming (which is not something I’ve heard of before!)

This is a really great suggestion, and definitely something I want to learn more about and utilize in the future. However, in this specific instance, I think I’d prefer to leave the code as-is. As you said, the current implementation is valid, and I can understand it readily-- I worry that, by changing it in line with what you wrote, I may have more difficulty maintaining it in the future (or, at least, until I understand functional programming better).
> Minor pfw_import() enhancement: read only the headers by using `header <- read.csv(file, nrows = 1)` instead of the entire file when checking colnames

Done!
> Cache the result of the imported data with memoise package

Right now, the results of pfw_import() are already cached as `.pfw_env$full_data` for use in pfw_zerofill()!
> Take the translation table out of the translation table subfolder and use `tools::R_user_dir("yourpkg", "data")` for update_taxonomy()

A very reasonable pair of suggestions; so reasonable, in fact, that I began changing the code to reflect them, only remembering why I’d made the subfolder in the first place about halfway through the process! Because the user can manually download the species translation table file from the website themselves (and may prefer to do so when annual taxonomy updates occur), they can name the file whatever they want. I found that having them put it in this subfolder regardless of the name helps me call it in the code.
> Consider breaking pfw_sitedata() down into separate functions (i.e. pfw_site_data_download(), pfw_sitedata_get_url(), pfw_sitedata_merge(), and with pfw_sitedata() as a wrapper to combine all the functions, like with pfw_filter())

While I understand how this could be functionally helpful, and maybe even clearer, I don’t see think it’s worthwhile because, unlike the sub-filtering functions compiled in pfw_filter(), a user wouldn’t really be using individual parts of pfw_sitedata() like that. They’d be importing the data and attaching it every time.
> Add an example for pfw_rollup() within pfw_filter() to show what it looks like with or without rollup for a given species (presumably one with both subspecies and slashes in the dataset)

Done!
> Consider adding a function to let users use month names/abbreviations instead of just numbers when calling pfw_date()

I considered this, and it is another good suggestion, but chose to continue supporting only numeric months to avoid further breaking of wrapped ranges (e.g., month = 11:2), which are already fairly finicky, and to avoid inconsistent parsing of character input.
> Fix month wrapping in pfw_date() so that it only wraps if a range is provided rather than just a pair of numbers (i.e. 11:2 wraps as c(11, 12, 1, 2), but c(11,2) does not wrap)

Done! I never expected the date selection to end up being one of the most frustrating and difficult to manage portions of this project...
> Add a bird name searching tool

While I see how this could be helpful, I am not sure it would make too much sense for someone to be using PFW and not know what species they are working with, even if it’s just the common name (which is why I allow common name, scientific name, or species code!). Maybe I’m misunderstanding, but I just don’t see a scenario where a user is working with this data but doesn’t know the bird names relevant to what they’re doing.
> Make the region lookup a packaged dataset rather than a .csv via usethis::use_data()

Done!
> Consider making the FeederWatch data dictionary an .rda file instead of a .csv

As you suggested, this is a manually-built file. Project FeederWatch does provide a data dictionary .xlsx file, but it has multiple sheets and isn’t formatted in a super R-readable way, so I manually assembled this one!

---
