### 2025-06-18 #698 @Visorbearer
Ah, I see what you're saying now, @paul-carteron-- apologies for misunderstanding before. I came up with a tweak for `pfw_species()` which would suggest the correct species if a spelling error is caught, though it would require adding `stringdist` as an import/dependency. What do you think?

```
  # Warn for unmatched species
  not_found <- species[!species %in% tolower(species_table$american_english_name) &
    !species %in% tolower(species_table$scientific_name) &
    !species %in% tolower(species_table$species_code)]

  if (length(not_found) > 0) {

    # Make a valid names list
    valid_names <- unique(c(
      species_table$american_english_name,
      species_table$scientific_name,
      species_table$species_code
    ))

    # Add a lowercase version for searching, keep title case for calling
    lower_valid <- tolower(valid_names)

    suggestions <- character()

    for (missing_sp in not_found) {
      distances <- stringdist::stringdist(missing_sp, lower_valid, method = "lv")
      min_dist <- min(distances)

      # Find the correct species
      if (min_dist <= 6) {
        best_match <- valid_names[which.min(distances)]
        suggestions <- c(suggestions, paste0("Did you mean: ", best_match, "?"))
      }
    }

    # Suggest the correct species if found
    if (length(suggestions) > 0) {
      warning("The following species were not found in the species table: ",
              paste(not_found, collapse = ", "), "\n",
      paste(suggestions, collapse = "\n"), call. = FALSE)
    }

    # If no similar names were found, don't suggest anything but still warn
    if (length(suggestions) == 0) {
      warning("The following species were not found in the species table: ",
              paste(not_found, collapse = ", "), call. = FALSE)
    }
  }
```
