### 2025-06-19 #698 @paul-carteron
This can surely help user !

However, I would have suggested multiple names like this :

```
  for (missing_sp in not_found) {
    distances <- stringdist::stringdist(missing_sp, lower_valid, method = "lv")

    best_matches <- valid_names[distances <= 3]
    suggestions <- c(suggestions,
                     paste0("Instead of `", missing_sp, "` did you mean one of: ",
                            paste0(best_matches, collapse = ", "), "?"))

  }
```
