### 2025-06-23 #698 @maelle
@paul-carteron could you please use the approval template https://devguide.ropensci.org/approval2template.html? Thank you for your continued feedback. 🙏

@SunnyTseng friendly reminder to have a look at @Visorbearer's response when you get the change. Thank you!
