### 2025-06-26 #698 @SunnyTseng
## Reviewer Response
Thanks @Visorbearer for the throughout response! In `update_taxanomy`, I see that your original code of having the dataset stored as `.csv` would make it easier to maintain and update the data. Thanks for the explaination. I learned a lot by reading your response and @paul-carteron 's review comments. Thank you, and congratulation again for the great package!

#### Final approval (post-review)

Estimated hours spent reviewing: 2 (second round review)
