### 2025-06-26 #698 @ropensci-review-bot
Approved! Thanks @Visorbearer for submitting and @SunnyTseng, @paul-carteron for your reviews! :grin:
